pyproject.toml
[project]
name = "voices"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alejandro Colomina", email = "ale@coolomina.dev"},
]
dependencies = [
    "Flask>=2.3.3",
    "Flask-CORS>=4.0.0",
    "Pillow>=10.0.1",
    "python-dotenv>=1.0.0",
    "Werkzeug>=2.3.0",
    "pydub>=0.25.1",
    "gunicorn>=21.2.0",
    "requests>=2.31.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-flask>=1.3.0",
]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "mypy>=1.16.0",
    "isort>=5.13.0",
    "pre-commit>=4.0.0",
    "bandit>=1.8.0",
    "types-Flask-Cors>=4.0.0",
]

[tool.pdm]
distribution = false

[tool.pdm.scripts]
start = "gunicorn -w 4 -b 0.0.0.0:5001 'app:create_app()'"
dev = "python app.py"
test = "pytest"
test-cov = "pytest --cov=. --cov-report=html --cov-report=term"
lint = "flake8 . --exclude=.venv,tests --extend-ignore=E203,W503,E501 --max-line-length=100"
format = "black ."
format-check = "black --check ."
isort = "isort ."
isort-check = "isort --check-only --diff ."
type-check = "mypy ."
pre-commit = "pre-commit run --all-files"
pre-commit-install = "pre-commit install"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | env
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".venv", "venv", "env"]
skip_glob = ["*/.venv/*", "*/venv/*", "*/env/*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    ".venv/",
    "venv/",
    "env/",
    "__pycache__/",
    "build/",
    "dist/",
    "tests/",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503", "E501"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    "build",
    "dist",
    ".eggs",
    "*.egg",
]
